+++ staticpipes/jinja2_environment.py
@@ -7,9 +7,10 @@ from staticpipes.source_directory import SourceDirectory
 
 class Jinja2Environment:
 
-    def __init__(self, autoescape=True):
+    def __init__(self, autoescape=True, filters: dict = {}):
         self._jinja2_environment: Optional[jinja2.Environment] = None
         self._autoescape = autoescape
+        self._filters = filters
 
     def get(self, source_directory: SourceDirectory) -> jinja2.Environment:
         if not self._jinja2_environment:
@@ -17,4 +18,7 @@ class Jinja2Environment:
                 loader=jinja2.FileSystemLoader(source_directory.dir),
                 autoescape=self._autoescape,
             )
+            for k, v in self._filters.items():
+                self._jinja2_environment.filters[k] = v
+
         return self._jinja2_environment