+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Added `filters` option to `Jinja2Environment` https://github.com/StaticPipes/StaticPipes/issues/29
+
 ## [0.2.0] - 2025-06-15
 
 ### Added